+++ TM1637.cpp
@@ -402,14 +402,14 @@ void TM1637::displayPChar( char * data )
 void TM1637::displayRaw(uint8_t * raw, uint8_t pointPos)
 {
   //  DEBUG
-  for (uint8_t d = 0; d < _digits; d++)
-  {
-    uint8_t x = raw[_digits - d - 1];
-    if (x < 0x10) Serial.print('0');
-    Serial.print(x, HEX);
-    Serial.print('-');
-  }
-  Serial.println();
+  // for (uint8_t d = 0; d < _digits; d++)
+  // {
+    // uint8_t x = raw[_digits - d - 1];
+    // if (x < 0x10) Serial.print('0');
+    // Serial.print(x, HEX);
+    // Serial.print('-');
+  // }
+  // Serial.println();
 
   uint8_t b = 0;
   _lastPointPos = pointPos;
@@ -461,9 +461,6 @@ void TM1637::dumpCache()
 
 void TM1637::hideSegment(uint8_t idx)
 {
-  Serial.println(__FUNCTION__);
-  Serial.println(idx);
-
   if (idx > 7) return;
   uint8_t tmp[8];
   for (int i = 0; i < 8; i++) tmp[i] = _data[i];
@@ -474,10 +471,6 @@ void TM1637::hideSegment(uint8_t idx)
 
 void TM1637::hideMultiSegment(uint8_t mask)
 {
-  //  debug info
-  Serial.println(__FUNCTION__);
-  Serial.println(mask, HEX);
-
   uint8_t tmp[8];
   for (int i = 0; i < 8; i++) 
   {