+++ TM1637.h
@@ -34,14 +34,20 @@ public:
   void displayFloat(float value, uint8_t fixedPoint);
   void displayHex(uint32_t value);
 
+  // DISPLAY FUNCTIONS META
   void displayClear();
   void displayRefresh();
-
+  //  EXPERIMENTAL 0.3.8
+  void hideSegment(uint8_t idx);
+  void hideMultiSegment(uint8_t mask);
+  
+  
   //  next 3 only tested on 4 digit display with colon
   void displayTime(uint8_t hh, uint8_t mm, bool colon);
   void displayTwoInt(int ll, int rr, bool colon = true);
+  //  Celsius  -9..99°C
   void displayCelsius(int temp, bool colon = false);
-
+  //  Fahrenheit -9..99°F
   void displayFahrenheit(int temp, bool colon = false);
 
 
@@ -78,9 +84,7 @@ public:
   void dumpCache();
 
 
-  //  EXPERIMENTAL
-  void hideSegment(uint8_t idx);
-  void hideMultiSegment(uint8_t mask);
+
 
 
 private: