+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.3.8] - 2023-07-15
 - fix #27
 - replaced function data\[8] with a class level data\[8].
-- add **hideSegment(n)**
-- add **displayFahrenheit()**
+- add **hideSegment(uint8_t n)** hide a single segment 
+- add **hideMultiSegment(uint8_t mask)** hide multiple segments
+- add **displayFahrenheit()** -9..99F
 - add **displayRefresh()** to refresh last written data.
 - update readme.md